move service-token implementation to browser/ and node/ folders
